TemplateDialog: keep private things private and header files simple
